Support non-hex strings in player colours, group sync-related methods
