+++ src/Component/Settings/Fields/IncludeTitle/IncludeTitle.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 
 namespace Beyondwords\Wordpress\Component\Settings\Fields\IncludeTitle;
 
-use Beyondwords\Wordpress\Component\Settings\Settings;
+use Beyondwords\Wordpress\Component\Settings\Sync;
 
 /**
  * IncludeTitle
@@ -44,7 +44,7 @@ class IncludeTitle
     {
         add_action('admin_init', array( $this, 'addSetting' ));
         add_action('pre_update_option_' . self::OPTION_NAME, function ($value) {
-            Settings::syncOptionToDashboard(self::OPTION_NAME);
+            Sync::syncOptionToDashboard(self::OPTION_NAME);
             return $value;
         });
         add_filter('option_' . self::OPTION_NAME, 'rest_sanitize_boolean');
