+++ src/Component/Settings/Fields/PlaybackFromSegments/PlaybackFromSegments.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 
 namespace Beyondwords\Wordpress\Component\Settings\Fields\PlaybackFromSegments;
 
-use Beyondwords\Wordpress\Component\Settings\Settings;
+use Beyondwords\Wordpress\Component\Settings\Sync;
 
 /**
  * PlaybackFromSegments
@@ -44,7 +44,7 @@ class PlaybackFromSegments
     {
         add_action('admin_init', array($this, 'addSetting'));
         add_action('pre_update_option_' . self::OPTION_NAME, function ($value) {
-            Settings::syncOptionToDashboard(self::OPTION_NAME);
+            Sync::syncOptionToDashboard(self::OPTION_NAME);
             return $value;
         });
         add_filter('option_' . self::OPTION_NAME, 'rest_sanitize_boolean');
