+++ src/Component/Settings/SettingsUtils.php
@@ -157,7 +157,6 @@ class SettingsUtils
      */
     public static function colorInput($label, $name, $value)
     {
-        $value = trim(ltrim($value, '#'));
         ?>
         <div class="color-input">
             <label>
@@ -165,15 +164,14 @@ class SettingsUtils
             </label>
             <output
                 for="<?php echo esc_attr($name); ?>"
-                style="background-color: #<?php echo esc_attr($value); ?>;"
+                style="background: <?php echo esc_attr($value); ?>; margin-right: 0.25rem;"
             ></output>
-            #
             <input
                 name="<?php echo esc_attr($name); ?>"
                 type="text"
                 value="<?php echo esc_attr($value); ?>"
                 class="small-text"
-                oninput="this.value = `${this.value}`.length > 1 ? `${this.value}`.replace(/[#]/g, '') : this.value; this.previousElementSibling.style.backgroundColor = 'transparent'; this.previousElementSibling.style.backgroundColor = `#${this.value}`"
+                oninput="this.previousElementSibling.style.background = 'transparent'; this.previousElementSibling.style.background = `${this.value}`"
             />
         </div>
         <?php
